xong cơ bản đồ án. người tìm việc còn trang cá nhân, tạo form thu thập thông tin gợi ý, nhà tuyển dụng còn chức năng thông báo khi có người ứng tuyển, admin xong quản lý xác thực và quản lý report

+++ backend/jobporal/src/main/java/com/project/jobportal/DTOs/InteractionDTO.java
@@ -12,9 +12,9 @@ public class InteractionDTO {
     private long jobId;
 
     @JsonProperty("is_save")
-    private int isSaved;
+    private Long isSaved;
 
     @JsonProperty("is_read")
-    private int isRead;
+    private Long isRead;
 
 }
